% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zeroclip.R
\name{zeroclipButton}
\alias{zeroclipButton}
\title{ZeroClipboard Button Output}
\usage{
zeroclipButton(inputId, label, clipText, icon = NULL, width = NULL)
}
\arguments{
\item{inputId}{The button Id.}

\item{label}{The  button Label.}

\item{clipText}{The text to be copied on click.}

\item{icon}{The button Id.}

\item{width}{The button width.}
}
\value{
htmlOutput
}
\description{
Create an Action Button with a ZeroClipboard data-clipboard-text attribute.
}

